Fix all audit bugs and warnings
Critical:
- InventoryBST: add destructor + deleteTree() to free all nodes on destroy
- InventoryBST/Stack/Queue: delete copy constructor and assignment operator
- main.cpp: guard input.empty() before input[0] on both command and play-again prompts
- HashTable::search: return Item by value (copy) instead of dangling pointer

Major:
- Room.cpp: fix pressEnter(bool) — only calls cin.ignore when a prior cin>> exists
  (ENEMY rooms pass true, ITEM rooms pass false — eliminates double-Enter bug)
- UI.cpp: display "?" for steps when pathfinder returns -1 (unreachable exit)
- Utils.h: move printBar() to shared inline header — removes duplication in UI/Room

Minor:
- Item/Enemy constructors: pass name by const reference instead of by value
- Sign-compare warnings fixed in main, HashTable, Sorter (cast to int)
- Map.cpp: exit cell always forced to EMPTY after random placement

+++ Sorter.cpp
@@ -22,7 +22,7 @@ void merge(std::vector<Item>& items, int left, int mid, int right) {
     int k = left;       // index into the original items vector
 
     // Compare front of each half and place the smaller one back
-    while (i < leftHalf.size() && j < rightHalf.size()) {
+    while (i < (int)leftHalf.size() && j < (int)rightHalf.size()) {
         if (leftHalf[i].getValue() <= rightHalf[j].getValue()) {
             items[k] = leftHalf[i];
             i++;
@@ -34,8 +34,8 @@ void merge(std::vector<Item>& items, int left, int mid, int right) {
     }
 
     // Copy any leftover elements from either half
-    while (i < leftHalf.size()) { items[k] = leftHalf[i]; i++; k++; }
-    while (j < rightHalf.size()) { items[k] = rightHalf[j]; j++; k++; }
+    while (i < (int)leftHalf.size()) { items[k] = leftHalf[i]; i++; k++; }
+    while (j < (int)rightHalf.size()) { items[k] = rightHalf[j]; j++; k++; }
 }
 
 // --- Merge Sort (recursive) ---
